Accomodate GitHub's token request requirements.
GitHub requires an 'accept: application/json' header, or it URL-encodes the token information, instead of returning it as JSON in the body.

GitHub encodes the returned 'scope' as a comma-separated string, instead of a list of strings.

This change always sends the 'accept' header, and adds the comma-separated list of scope names as a Json.Decode.oneOf option.

I tested that it works for AuthorizationCode grant flow on GitHub, Gmail, Facebook, and Gab.