remove 'authorization_code' example
In the end, there's not much reason why this example should exist.  The
code is there for completeness but obviously, it is wrong to use the
library as is in a public setup. When dealing with client applications,
the 'implicit' flow ought to be the way to go and therefore, it's better
to advertise this only.