Review examples and the way they're built
This to avoid having to carry unneeded dependencies in the library.
I use a symlink as a trick to make the module available to the examples
without having to require it explicitly as a dependency

This commit also makes 'deploy.sh' less verbose & fix example publication

- If the tag doesn't already exist, it displays an error message in the
console which we ought to ignore.

- Also, the publication to gh-pages was completely broken. It will now
nicely create a single commit with only the exported artefacts and get
back to master.

Finally, we are making examples a bit more sexy

- Review overall look-and-feel (colors, fonts, sizes, etc.)
- Remove the 'clientId' input and hard-code a demo clientId
- Display current 'Model' next to the actual content

- Write 'develop.sh' utility script to ease development